Update MeshletObject3D.ts
The bufferArray for the this.localPositionAttribute was too small because it didn't take into account that positions need a factor of 3 because of (x, y, z). When the number of meshlets is high, the color generator in the fragment shader only displays black when the meshlet value reaches a certain value. The new color generator also produces different colors for a very high number of meshlets.